More Changes for Final

+++ server/models/users.js
@@ -17,6 +17,14 @@ async function getAll(){
     return rows;
 }
 
+async function getAllUsersForAdmin(){
+    console.log("Called Get All For Admin")
+    const sql = `SELECT * FROM Users`;
+    const rows = await mysql.query(sql);
+    console.log(rows[0].id);
+    return rows;
+}
+
 async function getUserID(UserName, FirstName, LastName, DOB, Password){
     console.log("Getting User ID For Current Session")
     const rows = await mysql.query('Select id, FROM Users Where UserName = ? AND FirstName = ? And LastName = ? DOB = ? AND Password = ? ', [UserName, FirstName, LastName, DOB, Password])
@@ -59,6 +67,7 @@ async function login(email, password){
     const rows = await mysql.query(sql, [email]);
     if(!rows.length) throw { status: 404, message: "Sorry, that email address is not registered with us." };
     console.log(rows[0].Firstname);
+    console.log(rows[0].Password);
     console.log(Types.ADMIN);
     console.log(rows[0].User_Type)
     if(rows[0].User_Type == 5)
@@ -88,9 +97,31 @@ async function update(id, UserName, FirstName, LastName, DOB, Password, User_Typ
     return await mysql.query(sql, [params, id]);
 }
 
+async function ifadmin(id){
+    console.log('Checking Admin')
+    const sql = `Select * FROM Users WHERE id = ?`
+    const rows = mysql.query(sql, [id])
+    return rows;
+}
+
 async function remove(id){
-    const sql = `DELETE FROM Users WHERE id = ?`;
-    return await mysql.query(sql, [id]);
+    console.log('Removing User')
+    const admin = await ifadmin(id);
+    console.log('Outside admin')
+    console.log(admin[0].User_Type);
+    if(admin[0].User_Type == 5)
+      {
+         throw { status: 409, message: 'You cannot delete another admin' }
+      }
+    else{
+        const sql2 = `DELETE FROM ContactMethods WHERE User_id = ?`;
+        const rows2 = await mysql.query(sql2, [id])
+        const sql3 = `DELETE FROM Exercises WHERE User_id = ?`;
+        const rows3 = await mysql.query(sql3, [id])
+        const sql = `DELETE FROM Users WHERE id = ?`;
+        const rows = await mysql.query(sql, [id]);
+        return rows;
+    }
 }
 
 async function register(UserName, FirstName, LastName, DOB, Password, User_Type, email) {
@@ -106,25 +137,38 @@ async function register(UserName, FirstName, LastName, DOB, Password, User_Type,
     const user = await get(res.insertId);
     return user;
 }
-async function Add_A_New_Friend(Friends_URL_Page, Owner_id, Friend_id)
+
+async function getFriendlist(Owner_id){
+    console.log('getting list');
+    console.log(Owner_id);
+    //const sql = `SELECT *
+    //FROM Users U Join ContactMethods CM ON U.id=CM.User_id WHERE CM.Value=?`;
+    const sql = `SELECT * FROM Users U JOIN Friendlist FL ON U.ID = FL.Friends_id WHERE FL.Owner_id = ?`;
+    const rows = await mysql.query(sql, [Owner_id]);
+    if(!rows.length) throw {status: 404, message: "Sorry you have no friends added"};
+    return rows;
+}
+async function Add_A_New_Friend(Owner_id, Friend_id)
 {
+    console.log(Friend_id);
     if(Owner_id == Friend_id){
         throw { status: 409, message: 'You cannot add yourself' }
     }
-    const sql = `INSERT INTO Friendlist(created_at, Friends_URL_Page, Owner_id, Friend_id) VALUES ? ;`;
-    const params = [[new Date(), Friends_URL_Page, Owner_id, Friend_id]];
+    const sql = `INSERT INTO Friendlist(created_at, Friends_URL_Page, Owner_id, Friends_id) VALUES ? ;`;
+    const params = [[new Date(), Friends_URL_Page = null, Friend_id, Owner_id]];
     return await mysql.query(sql, [params]);
 }
 
 async function RemoveFriend(Owner_id, Friends_id)
 {
-    if(Owner_id == Friends_id){
-        throw { status: 409, message: 'You cannot delete yourself' }
-    }
-    const sql = `DELETE FROM Friendlist WHERE Friends_id = ?`;
-    return await mysql.query(sql, [id]);
+    console.log('deleting friend')
+    console.log(Owner_id)
+    console.log(Friends_id)
+    const sql = `DELETE FROM Friendlist WHERE id = ? AND Owner_id = ?`;
+    const rows = await mysql.query(sql, [Friends_id, Owner_id])
+    return rows;
 }
 
 const search = async q => await mysql.query(`SELECT id, UserName, FirstName, LastName FROM Users WHERE LastName LIKE ? OR FirstName LIKE ?; `, [`%${q}%`, `%${q}%`]);
 
-module.exports = { getAll, get, add, update, remove, getTypes, register, login, search, Types, Add_A_New_Friend, RemoveFriend, getUserFriends, getOwnerUserFriendId, getFriendUserId, getUserID}
+module.exports = { getAll, get, add, update, remove, getTypes, register, login, search, Types, Add_A_New_Friend, RemoveFriend, getUserFriends, getOwnerUserFriendId, getFriendUserId, getUserID, getFriendlist, ifadmin, getAllUsersForAdmin}
